+++ src/main/java/com/justifiedsolutions/justpdf/pdf/contents/SetRGBStrokeColor.java
@@ -7,18 +7,13 @@ package com.justifiedsolutions.justpdf.pdf.contents;
 
 import com.justifiedsolutions.justpdf.pdf.object.PDFReal;
 
-import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.Objects;
-
 /**
  * Implements the PDF command {@code RG} to set the stroke color in the RGB color space in a content stream.
  *
  * @see "ISO 32000-1:2008, 8.6.4.3"
  */
-public final class SetRGBStrokeColor implements ColorGraphicsOperator {
-    private final DeviceRGB colorSpace;
+public final class SetRGBStrokeColor extends DeviceRGBOperator {
+    private static final String OPERATOR_CODE = "RG";
 
     /**
      * Creates a new operator that sets the stroke color in the RGB color space in a content stream.
@@ -28,7 +23,7 @@ public final class SetRGBStrokeColor implements ColorGraphicsOperator {
      * @param blue  blue
      */
     public SetRGBStrokeColor(PDFReal red, PDFReal green, PDFReal blue) {
-        this(new DeviceRGB(red, green, blue));
+        super(new DeviceRGB(red, green, blue));
     }
 
     /**
@@ -37,43 +32,22 @@ public final class SetRGBStrokeColor implements ColorGraphicsOperator {
      * @param colorSpace the color space
      */
     public SetRGBStrokeColor(DeviceRGB colorSpace) {
-        this.colorSpace = Objects.requireNonNull(colorSpace);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(colorSpace);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        SetRGBStrokeColor that = (SetRGBStrokeColor) o;
-        return colorSpace.equals(that.colorSpace);
+        super(colorSpace);
     }
 
     @Override
     public boolean changesState(GraphicsState state) {
-        return !colorSpace.equals(state.getStrokeColorSpace());
+        return !getColorSpace().equals(state.getStrokeColorSpace());
     }
 
     @Override
     public void changeState(GraphicsState state) {
-        state.setStrokeColorSpace(colorSpace);
+        state.setStrokeColorSpace(getColorSpace());
     }
 
     @Override
-    public void writeToPDF(OutputStream pdf) throws IOException {
-        colorSpace.getRed().writeToPDF(pdf);
-        pdf.write(' ');
-        colorSpace.getGreen().writeToPDF(pdf);
-        pdf.write(' ');
-        colorSpace.getBlue().writeToPDF(pdf);
-        pdf.write(" RG\n".getBytes(StandardCharsets.US_ASCII));
+    protected String getOperatorCode() {
+        return OPERATOR_CODE;
     }
+
 }
